+++ src/components/concepts/Documentation.svelte
@@ -45,7 +45,6 @@
 
     interface Props {
         project: Project;
-
         collapse?: boolean;
     }